(|) scripts

+++ install.sh
@@ -1,4 +1,7 @@
-cp .env_example .env
+if [[ ! -f .env ]]; then
+    cp .env_example .env
+fi
+
 cd api
 sudo mysql < database.sql
 npm i
